Create Tweet_Intent_finder
I have created a script for the requirement to the extent of my knowledge which may or may not be absolute. I used some necessary modules for this analysis like nltk, MultinomialNB which is a naive bayes algorithm quiet commonly used for text analysis. I hope this script is better and needs some improvement.